backend/requirements.txt
alembic==1.14.1
asyncpg==0.30.0
chardet==5.2.0
fastapi==0.115.6
httpx==0.28.1
jellyfish==1.2.1
jinja2==3.1.5
openai==1.59.3
openpyxl==3.1.5
pandas==2.2.3
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.1
pytest-asyncio==1.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
sqlalchemy[asyncio]==2.0.48
uvicorn[standard]==0.34.0
email-validator==2.1.1
